--- src/main/java/ArgumentMustBeNumException.java
@@ -1,5 +1,0 @@
-public class ArgumentMustBeNumException extends DukeException {
-	public ArgumentMustBeNumException(String argumentName) {
-		super(String.format("The argument of %s must be a number.", argumentName));
-	}
-}
