--- src/main/java/DukeException.java
@@ -1,5 +1,0 @@
-public class DukeException extends RuntimeException {
-	public DukeException(String message) {
-		super(message);
-	}
-}
