--- src/main/java/DukeSideEffectException.java
@@ -1,5 +1,0 @@
-public class DukeSideEffectException extends RuntimeException {
-	public DukeSideEffectException(String message) {
-		super(message);
-	}
-}
