--- src/main/java/TaskListOutOfBoundsException.java
@@ -1,5 +1,0 @@
-public class TaskListOutOfBoundsException extends DukeException {
-	public TaskListOutOfBoundsException(int index) {
-		super(String.format("%d is out of bounds for the tasks list.", index));
-	}
-}
