--- src/main/java/UnidentifiedCommandException.java
@@ -1,5 +1,0 @@
-public class UnidentifiedCommandException extends DukeException {
-	public UnidentifiedCommandException() {
-		super("I'm sorry, but I don't know what that means :-(.");
-	}
-}
