+++ src/main/java/duke/exception/DukeSideEffectException.java
@@ -1,0 +1,7 @@
+package duke.exception;
+
+public class DukeSideEffectException extends RuntimeException {
+  public DukeSideEffectException(String message) {
+    super(message);
+  }
+}
