+++ src/main/java/duke/task/Todo.java
@@ -1,3 +1,5 @@
+package duke.task;
+
 public class Todo extends Task {
   public Todo(String description, boolean mark) {
     super(description, mark, 'T');
